4.15 4.23

+++ Part4/controllers/blogsRoute.js
@@ -1,12 +1,22 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blogModel");
-
+const User = require("../models/userModel");
+const jwt = require("jsonwebtoken");
 blogsRouter.get("/", async (request, response) => {
-	const blogs = await Blog.find({});
+	const blogs = await Blog.find({}).populate("user", {
+		username: 1,
+		id: 1,
+		name: 1,
+	});
 	response.status(200).json(blogs);
 });
 
 blogsRouter.post("/", async (request, response) => {
+	const user = request.user;
+
+	if (!user) {
+		return response.status(401).json({ error: "token missing or invalid" });
+	}
 	if (!request.body.title || !request.body.url) {
 		return response.status(400).end();
 	}
@@ -15,9 +25,12 @@ blogsRouter.post("/", async (request, response) => {
 		author: request.body.author,
 		url: request.body.url,
 		likes: request.body.likes ? request.body.likes : 0,
+		user: user.id,
 	});
 
 	const result = await blog.save();
+	user.blogs = user.blogs.concat(result._id);
+	await user.save();
 
 	response.status(201).json(result);
 });
@@ -30,8 +43,20 @@ blogsRouter.get("/:id", async (request, response) => {
 	response.status(200).json(blog);
 });
 blogsRouter.delete("/:id", async (request, response) => {
-	await Blog.findByIdAndDelete(request.params.id);
-	response.status(204).end();
+	const user = request.user;
+
+	if (!user) {
+		return response.status(401).json({ error: "token missing or invalid" });
+	}
+	const blog = await Blog.findById(request.params.id);
+	if (blog.user.toString() === request.user.id) {
+		await Blog.findByIdAndDelete(request.params.id);
+		response.status(204).end();
+	} else {
+		return response
+			.status(401)
+			.json({ error: "Unauthorized to delete the blog" });
+	}
 });
 blogsRouter.put("/:id", async (request, response) => {
 	const res = await Blog.findByIdAndUpdate(request.params.id, request.body, {
